# docs/openapi.json
{
  "components": {
    "schemas": {
      "Body_upload_documents_api_v1_rag_upload_post": {
        "properties": {
          "files": {
            "items": {
              "format": "binary",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "files"
        ],
        "title": "Body_upload_documents_api_v1_rag_upload_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "AI Real Estate Assistant API V4",
    "title": "AI Real Estate Assistant - Modern",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/rag/qa": {
      "post": {
        "description": "Simple QA over uploaded knowledge with citations.\nIf LLM is unavailable, returns concatenated context as answer.",
        "operationId": "rag_qa_api_v1_rag_qa_post",
        "parameters": [
          {
            "in": "query",
            "name": "question",
            "required": true,
            "schema": {
              "title": "Question",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "top_k",
            "required": false,
            "schema": {
              "default": 5,
              "title": "Top K",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Rag Qa",
        "tags": [
          "RAG"
        ]
      }
    },
    "/api/v1/rag/upload": {
      "post": {
        "description": "Upload text/markdown documents and index for local RAG (CE-safe).\nUnsupported types (pdf/docx) return a 422 in CE.",
        "operationId": "upload_documents_api_v1_rag_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_documents_api_v1_rag_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Upload Documents",
        "tags": [
          "RAG"
        ]
      }
    },
...
  }
}